Added an IIFE in recipeList with a two new internal functions and an updated forEach() function on the file scripts.js
If applied, this commit will have added the IIFE (Immediately Invoked Function Expression) a Functional Programming Paradigm
that organizes and protecting the global variables from being affected.
In order of that, a new variable named recipeRepository as been created with an attached function that contains the recipeList.
The functions 'add' and 'getAll' have been created inside of the new recipeRepository.
The function 'add' have an typeof that just allow object input.
A forEach() function was insert to update the previous 'for' Loop.