Added new 'addListItem' and 'showDetails' functions that generate and create a new Unordered List and is content. New typeof added in 'add' function and 'forEach' function improved.
If applied, this commit will have added a new 'addListItem' that creates a new recipeUnordered List, is own 'li' and a new 'button' that contains an eventListener with 'click' that will generate the new recipeList in the web page.
Another new 'showDetails' function have been created to be part of the eventListener function helping to generate further information.
The 'forEach()' function have been updated creating now a Loop in the function 'addListItem' that generate the recipeList.

Old information as 'filter()', 'check()' function or the previous conditions that we gave to the previous Loop being generated by the 'forEach()' function have been saved inside of /**/ in the end of the page